MuOxi is starting to look like something!
Still quite bare, but it is forming nicely. Right now the staging proxy server
works as intended, clients can connect to it via port :8000.

Tested on MushClient, lines sent back and forth between client and server is decode/encoded by
the LinesCodec struct.

Has ability to parse simple user commands, upon initial connection you will be presented by
a `create new` or `name` which should process the only two possibilities.

Still have to implement how information for each connect client will be stored, and how to load/save
that to disk or db.

A single struct acts as a shared resource between all connected clients with two fields, both a hashmap of:

`clients`
	- stores UID, (SocketAddr, Tx (mpsc::unboundedSender))

`accounts`
	- stores UID, ClientAccount (holds information about account of client itself)

The UID is new feature that was implemented, it is a `usize` data type and generated randomly when a client connects
this UID is given to all structs that hold information about the connected client, used for quickly identifiying client information between structs that cannot be safely shared between threads.

Right now Server struct and Client struct have to be seperated. Trying to have the server hold references to
the individual clients right now is proving to be difficult mainly because the Client struct holds tokio data structures
that represent TcpStream and mpsc::channels that do not inherit the `Copy` trait, therefore making it quite difficult

Using this method of a UID seems to be simple and effective.. but hey im still learning..

*sigh*... no one reads these...