After much research and consideration of how data will be stored within MuOxi, I have opted for a document-based database in MongoDB. There are pros and cons to using a SQL db, but I believe that by utilizing the flexibility of MongoDB it will allow easy changes to the data schema as time goes on.
A couple of new crates have been created. muoxi_db/ and muoxi_states.
The states.rs located within muoxi_staging will be removed as I feel a full fledged state machine at this
point in the development is pretty overkill. Maybe later, myself or someone else can implement a more robust way
of changing states.

Not completed yet, but much work as been done in creating simplified wrappers around the `mongodb` crate which is a
rust-driver for mongodb.

Two ways are used to essentially edit documents within MongoDB either via the use of `doc!` or constructing serializable structs
that are then parsed into BSONs to be consumed into MongoDB. I have opted for the latter that way, though not as strict as SQL
based databases, it will allow `some` structure to be implemented by creating structured that are going to be used to serialize/deserialze from the
database into the programs cache system.

The reason for this, is because MongoDB is extremely forgiving... it will allow duplicate documents (entities, rows, data) that consist of the exact same
fields.. the only thing different is the built in _id field that is generated for each `insert` of a document.

Right now a trait was created under muoxi_db/src/utils.rs that is a `MongoDocument` it essentially a helper trait to flag structs that are compatible with
the current Schema of MuOxi. In lamens terms, it essentially forces a user to have to essential fields for each struct

* uid
* name

both these fields are what is used when searching, updating, and inserting new documents into the database.

* uid was inspired by diku's vnum so that each document (record) is unique

Right now the database wrapper methods are not fully completed, and trial and error is being conducted to make sure they work as they should.

I feel at this point, getting a working API to the database is essential in order to continue on working on the network capabilities of both the staging/proxy server
and the actual game server

the benefit of using MongoDB is it allows for atomic operations so that both proxy/staging server AND game server can use the database without conflict..

yay...