A watchdog crate has been implemented, it's sole purpose is to monitor selected files (which will be the config/storage files) for MuOxi and upon a detected change, it will trigger a protocol that updates the MongoDB with the new information in the specific json files.
Right now, it is a naive implementation, as it completely rewrites the entire file to
database. This could potential slow down as time goes, on, but for now I have had to
physically implement delays in the code because rust is too damn fast, and the
monitoring is governed by the OS which doesn't pick up changes because rust writes to them too quickly.