Expierimenting with using Diesel ORM manager with PostgreSQL as backend.
Research about MongoDB and it seems to be too much trouble than it is worth,
Managed to create friendly wrappers and logic around searching, upserting, and deleting records.

For future, all structs that are serializable into Postgres and defined as tables are defined in their seperate files such as:
`clients.rs, accounts.rs, my_other_struct.rs`

within these files is another struct that represents all the methods and logic with interacting just with that table.

Still a work in progress, and I got like 2 hours of sleep tonight, so the code probably looks like mush (pun-intended)