Began implementing a command structure with associated commandsets. The design is still in its infancy, however the idea is as follows:
Depending on a variety of factors, including but not limited too: Connection State and Roles,

You have access to certain commands, if you type anything but those commands you will be
presented with a default 'huh' equivalent.

This inspiration comes from Evennia, mainly because I love the structure and the order of things.