Wow, what a ride...
Implemented a Command trait and CmdSet that holds a list of Cmds depending on a variety of factors.

What made this one a head scratcher, was that I began working with traits instead of objects themselves...

CmdSet.cmds is a vector not with an object type: T but rather a dynamically
allocated array of objects that share the same trait..

So CmdSet.cms == Vec<Box<dyn Command + Send>> the Send trait is to satisfy async capabilities across multiple threads.

This doesn't look like a lot fo code, but trying to figure it out initially was a pain in the ass, but the rewards outweigh the troubles.

Now when initializing a CmdSet you must first create appropriate commands using the Command trait, and the you can add the commands to a cmdset by the use of
smart pointers such as:

let mycmdset = CmdSet::new(vec![
				Box::new(CmdHelloWorld),
				Box::new(CmdLook),
				Box::new(CmdDrop),
			])

I will on a later date provide an easier macro to deal with this, and make it a bit easier on the eyes, in the mean time.
let's commit so I can go home.